move items n themes to common module
