+++ app/src/main/java/com/talhafaki/nextflixcomposable/ui/main/MainActivity.kt
@@ -5,7 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import com.talhafaki.nextflixcomposable.ui.theme.NextflixComposableTheme
+import com.talhafaki.common.theme.NextflixComposableTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
