+++ app/src/main/java/com/talhafaki/nextflixcomposable/ui/popular/PopularScreen.kt
@@ -11,9 +11,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
+import com.talhafaki.common.items.GridItem
+import com.talhafaki.common.theme.NextflixComposableTheme
 import com.talhafaki.domain.entity.NetworkMovie
-import com.talhafaki.nextflixcomposable.ui.common.GridItem
-import com.talhafaki.nextflixcomposable.ui.theme.NextflixComposableTheme
 
 /**
  * Created by tfakioglu on 12.December.2021
