--- app/src/main/res/drawable/ic_baseline_airplay.xml
@@ -1,6 +1,0 @@
-<vector android:height="24dp" android:tint="#C60404"
-    android:viewportHeight="24" android:viewportWidth="24"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M6,22l12,0l-6,-6z"/>
-    <path android:fillColor="@android:color/white" android:pathData="M21,3H3C1.9,3 1,3.9 1,5v12c0,1.1 0.9,2 2,2h4v-2H3V5h18v12h-4v2h4c1.1,0 2,-0.9 2,-2V5C23,3.9 22.1,3 21,3z"/>
-</vector>
