--- app/src/main/res/drawable/ic_baseline_trending_up.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#C60404"
-    android:viewportHeight="24" android:viewportWidth="24"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M16,6l2.29,2.29 -4.88,4.88 -4,-4L2,16.59 3.41,18l6,-6 4,4 6.3,-6.29L22,12V6z"/>
-</vector>
