+++ common/build.gradle
@@ -1,0 +1,49 @@
+plugins {
+    id 'com.android.library'
+    id 'kotlin-android'
+}
+
+android {
+    compileSdk 31
+
+    defaultConfig {
+        minSdk 21
+        targetSdk 31
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+    buildFeatures {
+        compose true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion Kotlin.compilerExtension
+        kotlinCompilerVersion Kotlin.compiler
+    }
+}
+
+dependencies {
+    implementation(project(":domain"))
+
+    implementation "androidx.compose.ui:ui:${AndroidX.compose}"
+    implementation "androidx.compose.material:material:${AndroidX.compose}"
+
+    // Coil
+    implementation "io.coil-kt:coil-compose:${CoilKt.coil}"
+}
